+++ src/dashboardWebView/components/Media/Pagination.tsx
@@ -1,9 +1,11 @@
+import { EventData } from '@estruyf/vscode';
 import { Messenger } from '@estruyf/vscode/dist/client';
 import { RefreshIcon } from '@heroicons/react/outline';
 import * as React from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
+import { DashboardCommand } from '../../DashboardCommand';
 import { DashboardMessage } from '../../DashboardMessage';
-import { LoadingAtom, MediaTotalSelector, PageAtom, SelectedMediaFolderSelector } from '../../state';
+import { LoadingAtom, MediaTotalSelector, PageAtom, SelectedMediaFolderSelector, SortingSelector } from '../../state';
 import { FolderCreation } from './FolderCreation';
 import { LIMIT } from './Media';
 import { PaginationButton } from './PaginationButton';
@@ -12,6 +14,7 @@ export interface IPaginationProps {}
 
 export const Pagination: React.FunctionComponent<IPaginationProps> = ({}: React.PropsWithChildren<IPaginationProps>) => {
   const selectedFolder = useRecoilValue(SelectedMediaFolderSelector);
+  const crntSorting = useRecoilValue(SortingSelector);
   const totalMedia = useRecoilValue(MediaTotalSelector);
   const [ , setLoading ] = useRecoilState(LoadingAtom);
   const [ page, setPage ] = useRecoilState(PageAtom);
@@ -46,11 +49,23 @@ export const Pagination: React.FunctionComponent<IPaginationProps> = ({}: React.
     Messenger.send(DashboardMessage.refreshMedia, { folder: selectedFolder });
   }
 
+  const mediaUpdate = (message: MessageEvent<EventData<{ key: string, value: any }>>) => {
+    if (message.data.command === DashboardCommand.mediaUpdate) {
+      setLoading(true);
+      Messenger.send(DashboardMessage.getMedia, {
+        page,
+        folder: selectedFolder || '',
+        sorting: crntSorting
+      });
+    }
+  }
+
   React.useEffect(() => {
     setLoading(true);
     Messenger.send(DashboardMessage.getMedia, {
       page,
-      folder: selectedFolder || ''
+      folder: selectedFolder || '',
+      sorting: crntSorting
     });
   }, [page]);
 
@@ -58,10 +73,24 @@ export const Pagination: React.FunctionComponent<IPaginationProps> = ({}: React.
     setLoading(true);
     Messenger.send(DashboardMessage.getMedia, {
       page: 0,
-      folder: selectedFolder || ''
+      folder: selectedFolder || '',
+      sorting: crntSorting
     });
     setPage(0);
   }, [selectedFolder]);
+
+  React.useEffect(() => {
+    setLoading(true);
+    Messenger.send(DashboardMessage.getMedia, {
+      page,
+      folder: selectedFolder || '',
+      sorting: crntSorting
+    });
+  }, [crntSorting]);
+
+  React.useEffect(() => {
+    Messenger.listen(mediaUpdate);
+  }, []);
 
   return (
     <nav