+++ src/helpers/CustomScript.ts
@@ -1,4 +1,4 @@
-import { CustomScript as ICustomScript } from '../models/PanelSettings';
+import { CustomScript as ICustomScript, ScriptType } from '../models/PanelSettings';
 import { window, env as vscodeEnv, ProgressLocation } from 'vscode';
 import { ArticleHelper } from '.';
 import { Folders } from '../commands/Folders';
@@ -8,21 +8,27 @@ import * as os from 'os';
 import { join } from 'path';
 import { Notifications } from './Notifications';
 import ContentProvider from '../providers/ContentProvider';
+import { Dashboard } from '../commands/Dashboard';
+import { DashboardCommand } from '../dashboardWebView/DashboardCommand';
 
 export class CustomScript {
 
-  public static async run(script: ICustomScript): Promise<void> {
+  public static async run(script: ICustomScript, path: string | null = null): Promise<void> {
     const wsFolder = Folders.getWorkspaceFolder();
 
     if (wsFolder) {
       const wsPath = wsFolder.fsPath;
 
-      if (script.bulk) {
-        // Run script on all files
-        CustomScript.bulkRun(wsPath, script);
+      if (script.type === ScriptType.MediaFile || script.type === ScriptType.MediaFolder) {
+        CustomScript.runMediaScript(wsPath, path, script);
       } else {
-        // Run script on current file.
-        CustomScript.singleRun(wsPath, script);
+        if (script.bulk) {
+          // Run script on all files
+          CustomScript.bulkRun(wsPath, script);
+        } else {
+          // Run script on current file.
+          CustomScript.singleRun(wsPath, script);
+        }
       }
     }
   }
@@ -79,6 +85,38 @@ export class CustomScript {
     });
   }
 
+  private static async runMediaScript(wsPath: string, path: string | null, script: ICustomScript): Promise<void>  {
+    if (!path) {
+      Notifications.error(`${script.title}: There was no folder or media path specified.`);
+      return;
+    }
+
+    return new Promise((resolve, reject) => {
+      window.withProgress({
+        location: ProgressLocation.Notification,
+        title: `Executing: ${script.title}`,
+        cancellable: false
+      }, async () => {
+        exec(`${script.nodeBin || "node"} ${join(wsPath, script.script)} "${wsPath}" "${path}"`, (error, stdout) => {
+          if (error) {
+            Notifications.error(`${script.title}: ${error.message}`);
+            resolve();
+            return;
+          }
+  
+          CustomScript.showOutput(stdout, script);
+  
+          Dashboard.postWebviewMessage({ 
+            command: DashboardCommand.mediaUpdate
+          });
+  
+          resolve();
+          return;
+        });
+      });
+    });
+  }
+
   private static async runScript(wsPath: string, article: matter.GrayMatterFile<string> | null, contentPath: string, script: ICustomScript): Promise<string | null> {
     return new Promise((resolve, reject) => {
       let articleData = "";
@@ -106,7 +144,7 @@ export class CustomScript {
       if (script.output === "editor") {
         ContentProvider.show(output, script.title, script.outputType || "text");
       } else {
-        window.showInformationMessage(`${script.title}: ${output}}`, 'Copy output').then(value => {
+        window.showInformationMessage(`${script.title}: ${output}`, 'Copy output').then(value => {
           if (value === 'Copy output') {
             vscodeEnv.clipboard.writeText(output);
           }