+++ src/helpers/FilesHelper.ts
@@ -1,15 +1,15 @@
-import * as vscode from 'vscode';
 import { Notifications } from './Notifications';
+import { Uri, workspace } from 'vscode';
 
 export class FilesHelper {
 
   /**
    * Retrieve all markdown files from the current project
    */
-  public static async getMdFiles(): Promise<vscode.Uri[] | null> {
-    const mdFiles = await vscode.workspace.findFiles('**/*.md', "**/node_modules/**,**/archetypes/**");
-    const markdownFiles = await vscode.workspace.findFiles('**/*.markdown', "**/node_modules/**,**/archetypes/**");
-    const mdxFiles = await vscode.workspace.findFiles('**/*.mdx', "**/node_modules/**,**/archetypes/**");
+  public static async getMdFiles(): Promise<Uri[] | null> {
+    const mdFiles = await workspace.findFiles('**/*.md', "**/node_modules/**,**/archetypes/**");
+    const markdownFiles = await workspace.findFiles('**/*.markdown', "**/node_modules/**,**/archetypes/**");
+    const mdxFiles = await workspace.findFiles('**/*.mdx', "**/node_modules/**,**/archetypes/**");
     if (!mdFiles && !markdownFiles) {
       Notifications.info(`No MD files found.`);
       return null;