+++ src/helpers/MediaLibrary.ts
@@ -1,3 +1,4 @@
+
 import { Dashboard } from '../commands/Dashboard';
 import { workspace } from 'vscode';
 import { JsonDB } from 'node-json-db/dist/JsonDB';