+++ src/models/PanelSettings.ts
@@ -1,4 +1,4 @@
-import { FileType } from "vscode";
+import { FileStat } from "vscode";
 import { DraftField } from ".";
 import { Choice } from "./Choice";
 import { DashboardData } from "./DashboardData";
@@ -9,6 +9,7 @@ export interface PanelSettings {
   tags: string[];
   date: DateInfo;
   categories: string[];
+  customTaxonomy: CustomTaxonomy[];
   freeform: boolean;
   scripts: CustomScript[];
   isInitialized: boolean;
@@ -32,12 +33,13 @@ export interface ContentType {
 export interface Field {
   title?: string;
   name: string;
-  type: "string" | "number" | "datetime" | "boolean" | "image" | "choice" | "tags" | "categories" | "draft";
+  type: "string" | "number" | "datetime" | "boolean" | "image" | "choice" | "tags" | "categories" | "draft" | "taxonomy";
   choices?: string[] | Choice[];
   single?: boolean;
   multiple?: boolean;
   isPreviewImage?: boolean;
   hidden?: boolean;
+  taxonomyId?: string;
 }
 
 export interface DateInfo {
@@ -63,11 +65,7 @@ export interface FolderInfo {
   lastModified: FileInfo[];
 }
 
-export interface FileInfo {
-  type: FileType;
-  ctime: number;
-  mtime: number;
-  size: number;
+export interface FileInfo extends FileStat {
   filePath: string;
   fileName: string;
 };
@@ -79,9 +77,21 @@ export interface CustomScript {
   bulk?: boolean;
   output?: "notification" | "editor";
   outputType?: string;
+  type?: ScriptType;
 }
 
 export interface PreviewSettings {
   host: string | undefined;
   pathname: string | undefined;
+}
+
+export interface CustomTaxonomy {
+  id: string;
+  options: string[];
+}
+
+export enum ScriptType {
+  Content = "content",
+  MediaFolder = "mediaFolder",
+  MediaFile = "mediaFile"
 }